Updated to v0.5.3
