+++ Cargo.toml
@@ -9,24 +9,12 @@ members = ["dnas/*/zomes/coordinator/*", "dnas/*/zomes/integrity/*", "crates/*"]
 resolver = "2"
 
 [workspace.dependencies]
-hdi = "0.6.0"
-hdk = "0.5.0"
+hdi = "0.6"
+hdk = "0.5"
+holochain_serialized_bytes = "*"
 serde = "1"
 
 [patch.crates-io]
-holochain = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-holochain_types = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-holochain_zome_types = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-holochain_integrity_types = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-holochain_p2p = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-holochain_timestamp = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-holochain_nonce = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-holochain_keystore = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-mr_bundle = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-holochain_conductor_api = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-holo_hash = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-fixt = { git = "https://github.com/guillemcordoba/holochain", branch = "develop-0.5" }
-
 kitsune2 = { git = "https://github.com/guillemcordoba/kitsune2", branch = "iroh-transport" }
 kitsune2_api = { git = "https://github.com/guillemcordoba/kitsune2", branch = "iroh-transport" }
 kitsune2_core = { git = "https://github.com/guillemcordoba/kitsune2", branch = "iroh-transport" }
