+++ crates/push_notifications_service_client/Cargo.toml
@@ -47,11 +47,3 @@ roles_types = { git = "https://github.com/darksoil-studio/roles-zome", branch = 
 
 tempdir = "0.3"
 colored = "2"
-
-[dev-dependencies]
-tempdir = "0.3.7"
-serde_yaml = "0.9"
-holo_hash = "0.5"
-fixt = "0.5"
-
-push-notifications-service-provider = { path = "../push_notifications_service_provider" }
