+++ crates/push_notifications_service_provider/tests/common.rs
@@ -2,6 +2,8 @@ use std::path::PathBuf;
 use std::{io::Write, time::Duration};
 
 use env_logger::Builder;
+use fixt::fixt;
+use holo_hash::fixt::AgentPubKeyFixturator;
 use holochain::prelude::{DnaModifiersOpt, RoleSettings, RoleSettingsMap, YamlProperties};
 use holochain_client::{AgentPubKey, AppWebsocket};
 use holochain_runtime::{vec_to_locked, HolochainRuntime, HolochainRuntimeConfig, NetworkConfig};
@@ -10,12 +12,6 @@ use push_notifications_service_provider::fcm_client::MockFcmClient;
 use push_notifications_service_provider::{read_from_file, run};
 use roles_types::Properties;
 use url2::url2;
-
-pub fn happ_developer_happ_path() -> PathBuf {
-    std::option_env!("HAPP_DEVELOPER_HAPP")
-        .expect("Failed to find HAPP_DEVELOPER_HAPP")
-        .into()
-}
 
 pub fn service_provider_happ_path() -> PathBuf {
     std::option_env!("SERVICE_PROVIDER_HAPP")
@@ -35,70 +31,7 @@ pub fn end_user_happ_path() -> PathBuf {
         .into()
 }
 
-pub async fn launch_infra_provider() -> (AppWebsocket, HolochainRuntime) {
-    let infra_provider = HolochainRuntime::launch(
-        vec_to_locked(vec![]),
-        HolochainRuntimeConfig::new(
-            tempdir::TempDir::new("test")
-                .expect("Could not make tempdir")
-                .into_path(),
-            network_config(),
-        ),
-    )
-    .await
-    .expect("Could not launch holochain runtime");
-
-    let admin_ws = infra_provider
-        .admin_websocket()
-        .await
-        .expect("Failed to connect AdminWebsocket");
-
-    let infra_provider_pub_key = admin_ws
-        .generate_agent_pub_key()
-        .await
-        .expect("Failed to generate pubkey");
-
-    let roles_properties = Properties {
-        progenitors: vec![infra_provider_pub_key.clone().into()],
-    };
-    let value = serde_yaml::to_value(roles_properties).unwrap();
-    let properties_bytes = YamlProperties::new(value);
-    let modifiers = DnaModifiersOpt {
-        properties: Some(properties_bytes),
-        ..Default::default()
-    };
-
-    let mut roles_settings = RoleSettingsMap::new();
-    roles_settings.insert(
-        String::from("push_notifications_service"),
-        RoleSettings::Provisioned {
-            membrane_proof: None,
-            modifiers: Some(modifiers),
-        },
-    );
-    let app_id = String::from("infra-provider");
-    let _app_info = infra_provider
-        .install_app(
-            app_id.clone(),
-            read_from_file(&client_happ_path())
-                .await
-                .expect("Failed to read infra provider happ"),
-            Some(roles_settings),
-            Some(infra_provider_pub_key.clone()),
-            None,
-        )
-        .await
-        .expect("Failed to install infra provider happ");
-
-    let app_ws = infra_provider
-        .app_websocket(app_id, holochain_client::AllowedOrigins::Any)
-        .await
-        .unwrap();
-
-    (app_ws, infra_provider)
-}
-
-fn network_config() -> NetworkConfig {
+pub fn network_config() -> NetworkConfig {
     let mut network_config = NetworkConfig::default();
     network_config.bootstrap_url = url2!("http://bad");
     network_config.signal_url = url2!("ws://bad");
@@ -106,9 +39,10 @@ fn network_config() -> NetworkConfig {
 }
 
 pub async fn launch(
-    infra_provider_pub_key: AgentPubKey,
+    progenitors: Vec<AgentPubKey>,
     roles: Vec<String>,
     happ_path: PathBuf,
+    network_seed: String,
 ) -> (AppWebsocket, HolochainRuntime) {
     let runtime = HolochainRuntime::launch(
         vec_to_locked(vec![]),
@@ -123,7 +57,7 @@ pub async fn launch(
     .expect("Could not launch holochain runtime");
 
     let roles_properties = Properties {
-        progenitors: vec![infra_provider_pub_key.clone().into()],
+        progenitors: progenitors.into_iter().map(Into::into).collect(),
     };
     let value = serde_yaml::to_value(roles_properties).unwrap();
     let properties_bytes = YamlProperties::new(value);
@@ -136,7 +70,7 @@ pub async fn launch(
                 membrane_proof: None,
                 modifiers: Some(DnaModifiersOpt {
                     properties: Some(properties_bytes.clone()),
-                    ..Default::default()
+                    network_seed: Some(network_seed.clone()),
                 }),
             },
         );
@@ -163,9 +97,8 @@ pub async fn launch(
 }
 
 pub struct Scenario {
-    pub infra_provider: (AppWebsocket, HolochainRuntime),
-    // pub service_provider: (AppWebsocket, HolochainRuntime),
-    pub happ_developer: (AppWebsocket, HolochainRuntime),
+    pub network_seed: String,
+    pub progenitors: Vec<AgentPubKey>,
     pub sender: (AppWebsocket, HolochainRuntime),
     pub recipient: (AppWebsocket, HolochainRuntime),
 }
@@ -181,8 +114,10 @@ pub async fn setup() -> Scenario {
         .filter_module("iroh", log::LevelFilter::Error)
         .init();
 
-    let infra_provider = launch_infra_provider().await;
-    let infra_provider_pubkey = infra_provider.0.my_pub_key.clone();
+    let network_seed = String::from("somesecret");
+    let progenitors = vec![fixt!(AgentPubKey)];
+
+    let p = progenitors.clone();
     tokio::spawn(async move {
         run::<MockFcmClient>(
             tempdir::TempDir::new("test")
@@ -191,13 +126,12 @@ pub async fn setup() -> Scenario {
             network_config(),
             String::from("test-app"),
             service_provider_happ_path(),
-            vec![infra_provider_pubkey],
+            p.clone(),
         )
         .await
         .unwrap();
     });
-
-    let infra_provider_pubkey = infra_provider.0.my_pub_key.clone();
+    let p = progenitors.clone();
     tokio::spawn(async move {
         run::<MockFcmClient>(
             tempdir::TempDir::new("test2")
@@ -206,36 +140,31 @@ pub async fn setup() -> Scenario {
             network_config(),
             String::from("test-app"),
             service_provider_happ_path(),
-            vec![infra_provider_pubkey],
+            p.clone(),
         )
         .await
         .unwrap();
     });
-    let happ_developer = launch(
-        infra_provider.0.my_pub_key.clone(),
-        vec![String::from("service_providers")],
-        happ_developer_happ_path(),
-    )
-    .await;
     let sender = launch(
-        infra_provider.0.my_pub_key.clone(),
+        progenitors.clone(),
         vec![String::from("service_providers")],
         end_user_happ_path(),
+        network_seed.clone(),
     )
     .await;
     let recipient = launch(
-        infra_provider.0.my_pub_key.clone(),
+        progenitors.clone(),
         vec![String::from("service_providers")],
         end_user_happ_path(),
+        network_seed.clone(),
     )
     .await;
 
     std::thread::sleep(Duration::from_secs(20));
 
     Scenario {
-        infra_provider,
-        // service_provider,
-        happ_developer,
+        network_seed,
+        progenitors,
         sender,
         recipient,
     }
