+++ crates/push_notifications_types/Cargo.toml
@@ -5,4 +5,5 @@ edition = "2021"
 
 [dependencies]
 hdi = { workspace = true }
+holochain_serialized_bytes = { workspace = true }
 serde = { workspace = true }
