+++ dnas/push_notifications_service/zomes/coordinator/push_notifications_service/Cargo.toml
@@ -10,6 +10,7 @@ name = "push_notifications_service"
 [dependencies]
 hdk = { workspace = true }
 
+holochain_serialized_bytes = { workspace = true }
 serde = { workspace = true }
 
 push_notifications_service_integrity = { path = "../../integrity/push_notifications_service" }
