+++ dnas/push_notifications_service/zomes/integrity/push_notifications_service/Cargo.toml
@@ -10,6 +10,7 @@ name = "push_notifications_service_integrity"
 [dependencies]
 hdi = { workspace = true }
 
+holochain_serialized_bytes = { workspace = true }
 serde = { workspace = true }
 
 push_notifications_types = { path = "../../../../../crates/push_notifications_types" }
