+++ dnas/service_providers/zomes/coordinator/push_notifications_gateway/Cargo.toml
@@ -10,6 +10,7 @@ name = "push_notifications_gateway"
 [dependencies]
 hdk = { workspace = true }
 
+holochain_serialized_bytes = { workspace = true }
 serde = { workspace = true }
 
 hc_zome_traits = { git = "https://github.com/holochain-open-dev/zome-traits", branch = "main" }
